moving TestContract to Integration Test
